Packages refactoring after moving Datastore from Service API to commons-dao
