+++ build.gradle
@@ -105,7 +105,7 @@ dependencies {
         compile 'com.epam.reportportal:commons-model:4.0.0'
         compile 'com.epam.reportportal:commons:3.1.0'
     } else {
-        compile('com.github.reportportal:commons-dao:ff8b4c1')
+        compile('com.github.reportportal:commons-dao:56ff0e4')
         compile 'com.github.reportportal:commons-rules:spb2-SNAPSHOT'
         compile 'com.github.reportportal:commons-model:3f17a8d'
         compile('com.github.reportportal:commons:71d4aca')
