+++ src/main/java/com/epam/ta/reportportal/store/service/DataStoreService.java
@@ -22,9 +22,9 @@ import com.epam.reportportal.commons.ContentTypeResolver;
 import com.epam.reportportal.commons.Thumbnailator;
 import com.epam.ta.reportportal.commons.BinaryDataMetaInfo;
 import com.epam.ta.reportportal.BinaryData;
-import com.epam.ta.reportportal.store.filesystem.DataEncoder;
-import com.epam.ta.reportportal.store.filesystem.DataStore;
-import com.epam.ta.reportportal.store.filesystem.FilePathGenerator;
+import com.epam.ta.reportportal.filesystem.DataEncoder;
+import com.epam.ta.reportportal.filesystem.DataStore;
+import com.epam.ta.reportportal.filesystem.FilePathGenerator;
 import com.google.common.base.Strings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
