+++ src/test/java/com/epam/ta/reportportal/store/service/DataStoreServiceTest.java
@@ -21,9 +21,9 @@ package com.epam.ta.reportportal.store.service;
 import com.epam.reportportal.commons.ContentTypeResolver;
 import com.epam.reportportal.commons.Thumbnailator;
 import com.epam.ta.reportportal.commons.BinaryDataMetaInfo;
-import com.epam.ta.reportportal.store.filesystem.DataEncoder;
-import com.epam.ta.reportportal.store.filesystem.DataStore;
-import com.epam.ta.reportportal.store.filesystem.FilePathGenerator;
+import com.epam.ta.reportportal.filesystem.DataEncoder;
+import com.epam.ta.reportportal.filesystem.DataStore;
+import com.epam.ta.reportportal.filesystem.FilePathGenerator;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
